# web/lib/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AapatMitra API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/tasks": {
   "get": {
    "tags": [
     "continuity"
    ],
    "summary": "List Tasks",
    "operationId": "list_tasks_api_v1_tasks_get",
    "parameters": [
     {
      "name": "assignee",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Assignee"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "open",
       "title": "Status"
      }
     },
     {
      "name": "dueBefore",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Duebefore"
      }
     },
     {
      "name": "villageId",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Villageid"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Tasks Api V1 Tasks Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "continuity"
    ],
    "summary": "Create Task",
    "operationId": "create_task_api_v1_tasks_post",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TaskIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FollowUpTaskOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tasks/{task_id}": {
   "patch": {
    "tags": [
     "continuity"
    ],
    "summary": "Patch Task",
    "operationId": "patch_task_api_v1_tasks__task_id__patch",
    "parameters": [
     {
      "name": "task_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Task Id"
      }
     },
     {
      "name": "if-match",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "additionalProperties": true,
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FollowUpTaskOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/tasks/{task_id}/commands": {
   "post": {
    "tags": [
     "continuity"
    ],
    "summary": "Task Command",
    "operationId": "task_command_api_v1_tasks__task_id__commands_post",
    "parameters": [
     {
      "name": "task_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Task Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TaskCommandIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FollowUpTaskOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/incentives/me": {
   "get": {
    "tags": [
     "continuity"
    ],
    "summary": "My Incentives",
    "operationId": "my_incentives_api_v1_incentives_me_get",
    "parameters": [
     {
      "name": "month",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Month"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response My Incentives Api V1 Incentives Me Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/leaderboards": {
   "get": {
    "tags": [
     "continuity"
    ],
    "summary": "Leaderboards",
    "operationId": "leaderboards_api_v1_leaderboards_get",
    "parameters": [
     {
      "name": "scope",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "village",
        "district"
       ],
       "type": "string",
       "default": "village",
       "title": "Scope"
      }
     },
     {
      "name": "scopeId",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Scopeid"
      }
     },
     {
      "name": "period",
      "in": "query",
      "required": false,
      "schema": {
       "const": "month",
       "type": "string",
       "default": "month",
       "title": "Period"
      }
     },
     {
      "name": "role",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "volunteer",
          "asha"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Role"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Leaderboards Api V1 Leaderboards Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "FollowUpTaskOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "patientId": {
      "type": "string",
      "format": "uuid",
      "title": "Patientid"
     },
     "patientName": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Patientname"
     },
     "patientShortCode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Patientshortcode"
     },
     "ashaId": {
      "type": "string",
      "format": "uuid",
      "title": "Ashaid"
     },
     "taskType": {
      "type": "string",
      "title": "Tasktype"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "priority": {
      "type": "string",
      "title": "Priority"
     },
     "sourceKind": {
      "type": "string",
      "title": "Sourcekind"
     },
     "sourceCaseId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sourcecaseid"
     },
     "dueDate": {
      "type": "string",
      "format": "date",
      "title": "Duedate"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "doneAt": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Doneat"
     },
     "doneEntryId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Doneentryid"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     }
    },
    "type": "object",
    "required": [
     "id",
     "patientId",
     "ashaId",
     "taskType",
     "priority",
     "sourceKind",
     "dueDate",
     "status",
     "version"
    ],
    "title": "FollowUpTaskOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "TaskCommandIn": {
    "properties": {
     "command": {
      "type": "string",
      "enum": [
       "Reopen",
       "Cancel",
       "MarkMissed"
      ],
      "title": "Command"
     },
     "args": {
      "additionalProperties": true,
      "type": "object",
      "title": "Args"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "command"
    ],
    "title": "TaskCommandIn"
   },
   "TaskIn": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "patientId": {
      "type": "string",
      "format": "uuid",
      "title": "Patientid"
     },
     "taskType": {
      "type": "string",
      "enum": [
       "anc_visit",
       "pnc_visit",
       "newborn_check",
       "bp_check",
       "sugar_check",
       "medicine_adherence",
       "referral_followup",
       "high_risk_recheck",
       "immunisation",
       "teleconsult_followup",
       "other"
      ],
      "title": "Tasktype"
     },
     "title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 120
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "dueDate": {
      "type": "string",
      "format": "date",
      "title": "Duedate"
     },
     "priority": {
      "type": "string",
      "enum": [
       "urgent",
       "normal"
      ],
      "title": "Priority",
      "default": "normal"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id",
     "patientId",
     "taskType",
     "dueDate"
    ],
    "title": "TaskIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}